+++ old-source/CheckList.Core/ClientApp/src/app/shared/models/actionitem.model.ts
@@ -1,8 +1,0 @@
-
-export class ActionItem {
-  listId: number;
-  actionId: number;
-  actionText: string;
-  categoryText: string;
-  isComplete: boolean;
-}