+++ old-source/CheckList.Core/ClientApp/src/app/shared/models/checkset.model.ts
@@ -1,5 +1,0 @@
-
-export class CheckSet {
-  setId: number;
-  setName: string;
-}