+++ old-source/CheckList.Core/ClientApp/src/app/shared/models/loginresult.model.ts
@@ -1,6 +1,0 @@
-
-export class LoginResult {
-  auth_token: string;
-  expires_in: number;
-  id: string;
-}